docker swarm, communication, fixes

+++ src/main/java/com/portfolio/showcase_spring/User/UserRepo.java
@@ -1,4 +1,4 @@
 package com.portfolio.showcase_spring.User;
 
 import org.springframework.data.repository.CrudRepository;
-public interface UserRepo extends CrudRepository<UserEntity, Integer> {}
+public interface UserRepo extends CrudRepository<UserEntity, Long> {}
